Make mypy work with and without pre-commit
The ternary expression with pytest.raises and nullcontext() does
not work with mypy since they are considered to be incompatible.
Therefore, we tell mypy to ignore this line.
However, in the pre-commit environment, mypy does not know about
pytest and black (from which nullcontext was imported),
so ignoring is not necessary here and yields another error.

To solve this discrepancy between standalone and pre-commit mypy,
we add pytest as dependency for pre-commit and avoid importing
from black by using the nullcontext from the standard lib.